# Fix stack overflow