--- a/ecmascript/codegen/src/expr.rs
+++ b/ecmascript/codegen/src/expr.rs
@@ -18,7 +18,7 @@ mod tests {
 
     #[test]
     fn bin_expr() {
-        assert_min("1+2+3+4+5", "1+2+3+4+5");
+        assert_min("1+2+3+4+5", "1+2+3+4+5;");
     }
 
     #[test]