--- a/ecmascript/codegen/src/lib.rs
+++ b/ecmascript/codegen/src/lib.rs
@@ -690,16 +690,46 @@ impl<'a> Emitter<'a> {
         emit!(node.right);
     }
 
-    #[emitter]
-    fn emit_bin_expr(&mut self, node: &BinExpr) -> Result {
-        self.emit_leading_comments_of_pos(node.span().lo())?;
-
+    /// Prints operator and right node of a binary expression.
+    #[inline(never)]
+    fn emit_bin_expr_trailing(&mut self, node: &BinExpr) -> Result {
         // let indent_before_op = needs_indention(node, &node.left, node.op);
         // let indent_after_op = needs_indention(node, node.op, &node.right);
         let need_space = match node.op {
             op!("in") | op!("instanceof") => true,
             _ => false,
         };
+
+        let need_pre_space = need_space
+            || match *node.left {
+                Expr::Update(UpdateExpr { prefix: false, .. }) => true,
+                _ => false,
+            };
+        if need_pre_space {
+            space!(self);
+        } else {
+            formatting_space!(self);
+        }
+        operator!(self, node.op.as_str());
+
+        let need_post_space = need_space
+            || match *node.right {
+                Expr::Unary(..) | Expr::Update(UpdateExpr { prefix: true, .. }) => true,
+                _ => false,
+            };
+        if need_post_space {
+            space!(self);
+        } else {
+            formatting_space!(self);
+        }
+        emit!(self, node.right);
+
+        Ok(())
+    }
+
+    #[emitter]
+    fn emit_bin_expr(&mut self, node: &BinExpr) -> Result {
+        self.emit_leading_comments_of_pos(node.span().lo())?;
 
         {
             let mut left = Some(node);
@@ -715,40 +745,20 @@ impl<'a> Emitter<'a> {
                 }
             }
 
+            let len = lefts.len();
+
             for (i, left) in lefts.into_iter().rev().enumerate() {
                 if i == 0 {
                     emit!(left.left);
                 }
-                operator!(left.op.as_str());
-                emit!(left.right);
+                // Check if it's last
+                if i + 1 != len {
+                    self.emit_bin_expr_trailing(left)?;
+                }
             }
         }
 
-        // emit!(node.left);
-
-        let need_pre_space = need_space
-            || match *node.left {
-                Expr::Update(UpdateExpr { prefix: false, .. }) => true,
-                _ => false,
-            };
-        if need_pre_space {
-            space!();
-        } else {
-            formatting_space!();
-        }
-        operator!(node.op.as_str());
-
-        let need_post_space = need_space
-            || match *node.right {
-                Expr::Unary(..) | Expr::Update(UpdateExpr { prefix: true, .. }) => true,
-                _ => false,
-            };
-        if need_post_space {
-            space!();
-        } else {
-            formatting_space!();
-        }
-        emit!(node.right);
+        self.emit_bin_expr_trailing(node)?;
     }
 
     #[emitter]