swc_ecma_codegen:
 - Fix codegen of large binary expressions.

swc_ecma_parser:
 - Fix stack overflow while parsing large binary expressions.

swc_ecma_transforms:
 - typescrip::strip: Fix stack overflow while handling large binary operations.
 - hygiene: Fix stack overflow by migrating it to `VisitMut`.
 - hygiene: Improve performance.
 - fixer: Fix stack overflow by migrating it to `VisitMut`.
 - resolver: Migrate to `VisitMut`.

---

# Benchmark

## master

```
(tf2) ➜  node-swc git:(master) ✗ node bench.js             
@swc/core x 269 ops/sec ±2.54% (78 runs sampled)
esbuild x 69.72 ops/sec ±2.01% (57 runs sampled)
typescript x 22.31 ops/sec ±8.98% (45 runs sampled)
babel x 21.00 ops/sec ±9.07% (41 runs sampled)
Transform rxjs/AjaxObservable.ts benchmark bench suite: Fastest is @swc/core

@swc/core x 173 ops/sec ±3.01% (65 runs sampled)
esbuild x 264 ops/sec ±1.29% (72 runs sampled)
Transform rxjs/AjaxObservable.ts async benchmark bench suite: Fastest is esbuild

@swc/core x 962 ops/sec ±1.23% (79 runs sampled)
esbuild x 1,242 ops/sec ±1.65% (79 runs sampled)
Transform rxjs/AjaxObservable.ts parallel benchmark bench suite: Fastest is esbuild

```

## After

```
(tf2) ➜  node-swc git:(deno-7279) ✗ node bench.js
@swc/core x 386 ops/sec ±0.26% (91 runs sampled)
esbuild x 52.41 ops/sec ±1.99% (54 runs sampled)
typescript x 23.62 ops/sec ±9.33% (47 runs sampled)
babel x 21.94 ops/sec ±8.34% (42 runs sampled)
Transform rxjs/AjaxObservable.ts benchmark bench suite: Fastest is @swc/core

@swc/core x 185 ops/sec ±2.96% (68 runs sampled)
esbuild x 258 ops/sec ±1.20% (75 runs sampled)
Transform rxjs/AjaxObservable.ts async benchmark bench suite: Fastest is esbuild

@swc/core x 1,066 ops/sec ±1.86% (77 runs sampled)
esbuild x 1,261 ops/sec ±1.08% (80 runs sampled)
Transform rxjs/AjaxObservable.ts parallel benchmark bench suite: Fastest is esbuild

```

## from rust
```
test full_es2016                  ... bench:   2,305,810 ns/iter (+/- 286,352)
```

1 / 2305810 * 1000000000 = 433.68 ops/sec... So napi binding is about 10% slower.

---

Related to https://github.com/denoland/deno/issues/7279